Update alphagateway_class.php
Lines 170 to 172 websites in languages other than english do refer to the gateway in greek language, for example a hotel customer goes to the page in german and he gets redirected to the bank's gateway in greek instead of in english. With the proposed change it would be the opposite.